services/task-service/requirements.txt
Flask==3.1.0
flask-cors==5.0.0
mysql-connector-python==9.1.0
PyJWT==2.10.1
python-dotenv==1.0.1
Werkzeug==3.1.3
Flask-Limiter==3.5.0
pydantic==2.6.1
